src/requirements.txt
pandas
numpy
openai
fastapi
redis
uvicorn
python-dotenv
sentence-transformers
# python-multipart